🧪 Integration test POC 🚀
This commit introduced a POC CI testing framework; this is not a
complete solution; it is a POC to show if we want to follow this way.

The main goal here is to use the kind cluster that GH actions are
running and use the Bats-core testing framework[0] to create edge devices
using an HTTP client(curl). Using yggradsil and device-worker will be
hard due to the nature of the workers, but we can try.

Each test file can contain a small test; the idea is to have a
lightweight CI on each PR, like:

- Creating an edgedevice and checking that all new MTLS workflow will
  work.
- Trigger a small Jmetter scenario that we can reuse, but with fewer
  clients.
- Run operator with race conditions and deadlocks checks
- Testing a deployment is correctly triggered, etc..
- Be able to test getting starting guides.

I know it's bash code, but it is easy to understand and keep up to date.
(I hope I do not regret about this)

[0] https://bats-core.readthedocs.io/en/stable/tutorial.html#your-first-test

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>