Set limits and concurrency for k4e controller
In order to prevent from having an impact on the cluster, setting limits
for the k4e controller.

In addition, to save memory, running the reconcilers with a higher
concurrency consume less memory than increasing number of replicas.

Memory and CPU request and limits need to be tuned based on expected
amount of devices.

Signed-off-by: Moti Asayag <masayag@redhat.com>