Auth: Registration with Certicate signed request
This commit adds new parameters to the "registration" directive, where a user
will send a CSR to the endpoint, the operator will ask CA provider to get it
signed and return the certificate to the user.

At the moment, CSR is not required, and the main reason is to not break the
clients right now, but a follow-up commit will fix that and it'll implement
Authz and the missing tests.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>