e2e: Fixes and gather logs
Podman was not started when running the edgedevice, so the e2e tests
were failing because can't deploy to the edgedevice.

Also, I added a way to dump logs, ATM in the console, but in the next
days I'll evolve to a file, so logs can be attached to the end of the
job. For making it faster, logs are only retrieved when the test fails
and on the JustAfterEach function, so the device and containers, are
still in there.

Fix: https://issues.redhat.com/browse/ECOPROJECT-649

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>