Makefile: add GINKGO_OPTIONS to test-fast
When using `make test-fast TEST_PACKAGES="./controllers/"` at the moment
was not possible to run any focus test, something that it's super useful
to run it.

With this change, `--skip e2e` is removed from test-fast and a new
option GINKGO_OPTIONS added to test-fast target. When the make test is
called it'll skip the 2e2 test.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>