Registration: set correct namespace
When signing the CSR given by the device, it signs correctly, but the
namespace set on the function was `h.initialNamesapce`. When device try
to retrieve the config,  it retuns 404 that device cannot be found,
because the certificate does not contain the valid metadata.

This commits change the namespace to the correct one, and adds a
specific check on the unittest to avoid this kind of issues in future.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>