Reused target for pre-build actions
The pre-build actions are grouped into a single target `pre-build`
and reused when needed to produce the same output when building
the project.

In addition, `fast-build` executes minimal set of targets to build the
manager binary. Before submitting the code, the developer will have to
run `make all` to assure code is formatted, vetted and organized as
expected.

Signed-off-by: Moti Asayag <masayag@redhat.com>