CI: Fix some invalid typos
PR #275 was a test PR to debug why the 2e2 testing CI job was failing
because I was unable to replicate. That PR had a cherry-pick of PR #267,
and some other code to help to understand why it didn't work.

With this change, some typos are deleted and also some invalid commands
made for debugging.

And yes, the commits on that PR were super bad, I know, I didn't expect
that got merged in any way, sorry!

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>